Add a clear cache button
Invalidates the entire cache allowing data reloading.

This is a not-amazing fix for any sync issues that may arise, in particular as a result of the fact that the protobuf deserializer we use does not validate data passed to it, occasionally resulting in junk data being stored to the cache.